feat: Implement global alert utility and update script loading order
