+++ src/scripts/eventHandlers.js
@@ -123,62 +123,14 @@ document.getElementById('geneAnnotationBtn').addEventListener('click', async(e) 
 });
 
 /**
- * Display an alert message to the user
- * @param {string} message - The message to display
- * @param {string} type - The type of alert: 'info', 'success', 'warning', or 'error'
- * @param {number} timeout - Time in milliseconds before the alert disappears (0 for no auto-dismiss)
- */
-function showAlert(message, type = 'info', timeout = 0) {
-    // Map alert types to Bootstrap classes and icons
-    const typeMap = {
-        'info': { class: 'alert-info', icon: 'bi-info-circle' },
-        'success': { class: 'alert-success', icon: 'bi-check-circle' },
-        'warning': { class: 'alert-warning', icon: 'bi-exclamation-triangle' },
-        'error': { class: 'alert-danger', icon: 'bi-exclamation-circle' }
-    };
-
-    // Get the appropriate class and icon
-    const alertClass = typeMap[type] ? .class || 'alert-info';
-    const alertIcon = typeMap[type] ? .icon || 'bi-info-circle';
-
-    // Create alert element
-    const alertDiv = document.createElement('div');
-    alertDiv.className = `alert ${alertClass} alert-dismissible fade show position-fixed`;
-    alertDiv.style.bottom = '20px';
-    alertDiv.style.left = '20px';
-    alertDiv.style.zIndex = '9999';
-    alertDiv.style.maxWidth = '400px';
-
-    // Add content with icon
-    alertDiv.innerHTML = `
-        <i class="bi ${alertIcon} me-2"></i>
-        ${message}
-        <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-    `;
-
-    // Add to document
-    document.body.appendChild(alertDiv);
-
-    // Auto-dismiss if timeout is provided
-    if (timeout > 0) {
-        setTimeout(() => {
-            alertDiv.classList.remove('show');
-            setTimeout(() => alertDiv.remove(), 300); // Remove after fade animation
-        }, timeout);
-    }
-
-    return alertDiv;
-}
-
-/**
  * Handle direct gene annotation when modal is unavailable
  */
 async function handleGeneAnnotation() {
     try {
         console.log("Starting gene annotation process directly...");
 
-        // Show loading alert
-        showAlert("Loading gene annotations...", "info");
+        // Show loading alert - use the global function instead of imported one
+        window.showAlert("Loading gene annotations...", "info");
 
         // Try loading the gene data 
         const loaded = await loadGeneData();
@@ -188,14 +140,14 @@ async function handleGeneAnnotation() {
             // Annotate all nodes in the graph
             const annotatedCount = await annotateAllNodes(STATE.cy);
 
-            // Show success alert with auto-dismiss after 3 seconds
-            showAlert(`Annotated ${annotatedCount} nodes with gene information!`, "success", 3000);
+            // Show success alert with auto-dismiss after 3 seconds - use global function
+            window.showAlert(`Annotated ${annotatedCount} nodes with gene information!`, "success", 3000);
         } else {
             console.error("Could not load gene data or graph not initialized");
-            showAlert("Failed to load gene annotations.", "error");
+            window.showAlert("Failed to load gene annotations.", "error");
         }
     } catch (error) {
         console.error("Error in gene annotation:", error);
-        showAlert("Error in gene annotation process.", "error");
+        window.showAlert("Error in gene annotation process.", "error");
     }
 }
